+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/resetter/DubboResetter.java
@@ -138,6 +138,18 @@ public class DubboResetter extends AbstractResetter {
          */
         @SuppressWarnings("SameReturnValue")
         boolean resetJob() throws ServiceException;
+
+        /**
+         * 重置主管功能。
+         *
+         * <p>
+         * 因为 Dubbo 广播响应机制无法处理 void 返回类型，所以方法需要返回一个结果。
+         *
+         * @return 恒为 true。
+         * @throws ServiceException 服务异常。
+         */
+        @SuppressWarnings("SameReturnValue")
+        boolean resetSupervise() throws ServiceException;
     }
 
     @org.springframework.stereotype.Service
@@ -154,6 +166,17 @@ public class DubboResetter extends AbstractResetter {
             try {
                 LOGGER.info("接收到作业功能重置消息, 正在重置作业功能...");
                 context.resetJob();
+                return true;
+            } catch (Exception e) {
+                throw ServiceExceptionHelper.logParse("发生异常", LogLevel.WARN, e, sem);
+            }
+        }
+
+        @Override
+        public boolean resetSupervise() throws ServiceException {
+            try {
+                LOGGER.info("接收到主管功能重置消息, 正在重置主管功能...");
+                context.resetSupervise();
                 return true;
             } catch (Exception e) {
                 throw ServiceExceptionHelper.logParse("发生异常", LogLevel.WARN, e, sem);