+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/resetter/FixedRateResetter.java
@@ -58,6 +58,8 @@ public class FixedRateResetter extends AbstractResetter {
                 LOGGER.info("计划时间已到, 重置相关功能...");
                 LOGGER.info("重置作业功能...");
                 context.resetJob();
+                LOGGER.info("重置主管功能...");
+                context.resetSupervise();
             } catch (Exception e) {
                 String message = "重置器 " + FixedRateResetter.this +
                         " 执行重置调度时发生异常, 相关工功能可能不会完全重置, 异常信息如下: ";